✨ Move Server Side utils to separate directory
Because I'm importing from an index file instead of a specific
file, when I use a utility function on the client, it's trying to
pull in the entire directory. This is creating problems.

This commit is about separating out utility functions that rely on
node modules (e.g., `fs`) and which will break if invoked on the
client.