🐛 Fix Search
Instead of building the search indexes in the build script, this
commit moves that the compilation of the index to the first time
the search api is hit.

This has the benefit of avoiding the export function of flexsearch
which clearly doesn't work the way it's advertised (or at least I
didn't figure it out!).