✨ Add pills (#25)
* ✨ add tags and categories to slugDictionary

* ✨ Add pills to search results

* ✨ Add pills to blog index

* ✨ Add some links and adjust styles

* 🔨 update content & add dayjs as dev dep

dayjs is needed for the scripts included in the submodules as of a
recent update there. to solve the problem where the build process
wasn't able to find it, I added this as a dependency.

* 🧰 Fix dayjs dependency

Instead of adding dayjs as a dependency, we are able to simply
exclude the scripts from the submodule, where this is required.

* 🐛 Fix Search

Instead of building the search indexes in the build script, this
commit moves that the compilation of the index to the first time
the search api is hit.

This has the benefit of avoiding the export function of flexsearch
which clearly doesn't work the way it's advertised (or at least I
didn't figure it out!).