✨ Add TW spacing package + fix layout styling
Spacing package adds the lobotomized owl (*+*) selector to the
utility classes of tailwind see the package tailwindcss for more
detail.

Also found a bug in the tailwind.config that had previously
excluded the layouts from getting styles applied during JIT mode
(as the directory was not included in the content key).