✨ Augment Frontmatter type with backlink & pinned
the backlink key on frontmatter is what will allow notes to be
tied together through references.

the pinned is an experiment that might allow me to have a section
for pinned/favorited notes to make them easier to access.

I think this will be especially relevant as the number of notes I'm
dealing with grows.

Eventually it'll be useful to have them in a completely different
section perhaps, but right now, I expect they'll just be sorted
to the front of the list.